Resuelve los conflictos entre las dos ramas de trabajo en paralelo:
consolida en PlantillaService (elimina Plantillas), en Descarga para
las cabeceras Content-Disposition (NombresDeArchivo pasa a usarse solo
para sanitizar nombres de entradas de ZIP), conserva ambas paginas
legales (/privacidad + /terminos y /terminos-y-condiciones) en
SecurityConfig, combina el cooldown de reenvio de codigo con la logica
de reusar el codigo vigente en OlvideContrasenaController, y mantiene
el manejo de usuario borrado a mitad de sesion en CambiarContrasenaController.

De paso corrige un bug ya presente en main: requiresChannel() quedo
roto en Spring Security 7.1.0 porque el JAR de spring-security-web ya
no trae ChannelDecisionManager (referenciada igual desde
spring-security-config), lo que tiraba NoClassDefFoundError al levantar
el contexto. Se reemplaza por redirectToHttps(), el reemplazo soportado
para forzar HTTPS.

+++ src/main/java/com/EscuelaEmpresa/gestor_pasantes/dto/DiaForm.java
@@ -1,5 +1,6 @@
 package com.EscuelaEmpresa.gestor_pasantes.dto;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 import org.springframework.format.annotation.DateTimeFormat;
@@ -12,7 +13,9 @@ public class DiaForm {
     private LocalDate fecha;
 
     private String descripcion;
-    private Float horas;
+    // BigDecimal y no Float: las horas se suman, y en binario 7.5 + 8.25 no da
+    // 15.75 exacto. Con dos decimales fijos la cuenta cierra siempre.
+    private BigDecimal horas;
 
     public DiaForm() {}
 
@@ -29,6 +32,6 @@ public class DiaForm {
     public String getDescripcion() { return descripcion; }
     public void setDescripcion(String descripcion) { this.descripcion = descripcion; }
 
-    public Float getHoras() { return horas; }
-    public void setHoras(Float horas) { this.horas = horas; }
+    public BigDecimal getHoras() { return horas; }
+    public void setHoras(BigDecimal horas) { this.horas = horas; }
 }
